requirements.txt
# Core Framework
fastapi
uvicorn

# Database and Environment
pymongo
python-dotenv

# Security (JWT, Hashing, etc.)
passlib[bcrypt]
python-jose[cryptography]

# Data Validation
pydantic